backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
beautifulsoup4==4.12.3
bleach==6.1.0
redis==5.0.7
rq==1.16.2
supabase==2.5.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
httpx==0.27.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
email-validator==2.1.1
